애니메이션 길이를 인스펙터의 클립들로 계산

+++ sync_motion/Assets/1.Scripts/PlayerController.cs
@@ -108,7 +108,7 @@ public class PlayerController : MonoBehaviour, IPunObservable
     {
         if (targetHor == 0f && targetVer == 0f)
         {
-            animationController.PlayDance(value.Get<float>());
+            animationController.PlayDance((int)value.Get<float>());
         }
     }
